[update]整数になる部分はunsigendを付与

+++ server/database/migrations/2020_05_28_234751_create_users_table.php
@@ -26,11 +26,11 @@ class CreateUsersTable extends Migration
             $table->string('sei_kana', 100)->comment('セイ');
             $table->string('mei_kana', 100)->comment('メイ');
             $table->string('img')->nullable()->comment('画像のS3パス');
-            $table->tinyInteger('gender')->nullable()->comment('性別(1:男 2:女)');
+            $table->unsignedSmallInteger('gender')->nullable()->comment('性別(1:男 2:女)');
             $table->date('birth')->nullable()->comment('誕生日');
-            $table->tinyInteger('role')->comment('権限(1:システム管理者  3:管理者 5:共有アカウント)');
+            $table->unsignedTinyInteger('role')->comment('権限(1:システム管理者  3:管理者 5:共有アカウント)');
             $table->string('password', 255)->nullable()->comment('パスワード');
-            $table->tinyInteger('status')->comment('状態(1:会員 2:退会)');
+            $table->unsignedSmallInteger('status')->comment('状態(1:会員 2:退会)');
             $table->timestamp('email_verified_at')->nullable();
             $table->rememberToken();
             $table->softDeletes();
